[MCUX-83760] Fix MISRA C-2012 Rule 10.4
Essential type of the left hand operand element_size (signed) is not the same as that of the right operand sizeof (uint32_t)(unsigned).
Fix by casting element_size to uint32_t.

+++ lib/rpmsg_lite/porting/environment/rpmsg_env_threadx.c
@@ -656,7 +656,7 @@ int32_t env_create_queue(void **queue, int32_t length, int32_t element_size)
         return -1;
     }
 
-    if (TX_SUCCESS == tx_queue_create((TX_QUEUE *)queue_ptr, NULL, (element_size / RL_WORD_SIZE),
+    if (TX_SUCCESS == tx_queue_create((TX_QUEUE *)queue_ptr, NULL, ((uint32_t)element_size / RL_WORD_SIZE),
                                       (VOID *)msgq_buffer_ptr, (length * element_size)))
     {
         *queue = (void *)queue_ptr;
